fix: update files to resolve PR comments
