+++ .eslintrc
@@ -1,7 +1,4 @@
 {
     "extends": ["xo", "xo-typescript"],
-    "parser": "@typescript-eslint/parser",
-    "rules": {
-        "@typescript-eslint/no-dynamic-delete":"off"
-    }
+    "parser": "@typescript-eslint/parser"
 }
