+++ package.json
@@ -4,7 +4,7 @@
 	"description": "Simple config handling for your app or module",
 	"license": "MIT",
 	"repository": "sindresorhus/conf",
-	"main": "dist/index.js",
+	"main": "dist/source",
 	"funding": "https://github.com/sponsors/sindresorhus",
 	"author": {
 		"name": "Sindre Sorhus",
@@ -15,14 +15,15 @@
 		"node": ">=8"
 	},
 	"scripts": {
+		"test": "xo && tsc --noEmit && nyc ava",
+		"prepare": "npm run build",
 		"build": "del dist && tsc",
-		"test": "npm run build && xo source/** && nyc ava",
 		"prepublishOnly": "npm run build"
 	},
 	"files": [
-		"dist"
+		"dist/source"
 	],
-	"types": "dist",
+	"types": "dist/source",
 	"keywords": [
 		"config",
 		"store",
@@ -43,64 +44,62 @@
 		"cache"
 	],
 	"ava": {
-		"babel": false,
-		"compileEnhancements": false,
-		"extensions": [
-			"ts"
+		"files": [
+			"test/*"
 		],
-		"require": [
-			"ts-node/register"
-		]
+		"timeout": "1m",
+		"typescript": {
+			"rewritePaths": {
+				"test/": "dist/test/"
+			}
+		}
 	},
 	"xo": {
 		"extends": "xo-typescript",
 		"extensions": [
 			"ts"
-		],
-		"rules": {
-			"@typescript-eslint/no-dynamic-delete": "off"
-		}
+		]
 	},
 	"nyc": {
 		"extension": [
-			".ts"
+			"*.ts"
 		],
-		"include": [
-			"source/**"
+		"exclude": [
+			"**/test/**"
 		]
 	},
 	"dependencies": {
-		"ajv": "^6.10.2",
-		"crypto": "^1.0.1",
-		"debounce-fn": "^3.0.1",
+		"ajv": "^6.12.0",
+		"debounce-fn": "^4.0.0",
 		"dot-prop": "^5.2.0",
 		"env-paths": "^2.2.0",
-		"json-schema-typed": "^7.0.2",
-		"make-dir": "^3.0.0",
+		"json-schema-typed": "^7.0.3",
+		"make-dir": "^3.0.2",
 		"onetime": "^5.1.0",
-		"pkg-up": "^3.0.1",
-		"semver": "^6.2.0",
-		"write-file-atomic": "^3.0.1"
+		"pkg-up": "^3.1.0",
+		"semver": "^7.1.3",
+		"write-file-atomic": "^3.0.3"
 	},
 	"devDependencies": {
-		"@sindresorhus/tsconfig": "^0.6.0",
+		"@ava/typescript": "^1.1.1",
+		"@sindresorhus/tsconfig": "^0.7.0",
 		"@types/node": "^12.12.12",
-		"@types/semver": "^6.2.0",
-		"@types/write-file-atomic": "^2.1.2",
-		"@typescript-eslint/eslint-plugin": "^2.9.0",
-		"@typescript-eslint/parser": "^2.9.0",
-		"ava": "^2.4.0",
-		"clear-module": "^4.0.0",
+		"@types/semver": "^7.1.0",
+		"@types/write-file-atomic": "^3.0.0",
+		"@typescript-eslint/eslint-plugin": "^2.21.0",
+		"@typescript-eslint/parser": "^2.21.0",
+		"ava": "^3.4.0",
+		"clear-module": "^4.1.0",
 		"del": "^5.1.0",
 		"del-cli": "^3.0.0",
 		"delay": "^4.3.0",
-		"eslint-config-xo-typescript": "^0.22.0",
-		"nyc": "^14.1.1",
+		"eslint-config-xo-typescript": "^0.26.0",
+		"nyc": "^15.0.0",
 		"p-event": "^4.1.0",
-		"tempy": "^0.3.0",
-		"ts-node": "^8.5.2",
+		"tempy": "^0.4.0",
+		"ts-node": "^8.6.2",
 		"tsd": "^0.11.0",
-		"typescript": "^3.7.5",
-		"xo": "^0.25.3"
+		"typescript": "^3.8.2",
+		"xo": "^0.26.1"
 	}
 }
