+++ test/index.ts
@@ -1,4 +1,4 @@
-/* eslint-disable no-new, @typescript-eslint/no-empty-function, ava/use-test */
+/* eslint-disable no-new, @typescript-eslint/no-empty-function */
 import fs = require('fs');
 import path = require('path');
 import tempy = require('tempy');
@@ -7,9 +7,10 @@ import pkgUp = require('pkg-up');
 import clearModule = require('clear-module');
 import pEvent = require('p-event');
 import delay = require('delay');
-import anyTest, { TestInterface } from 'ava';
-import Conf from '../source';
-import { JSONSchemaType } from 'json-schema-typed';
+import anyTest, {TestInterface} from 'ava';
+import {version as packageVersion} from '../package.json';
+// eslint-disable-next-line import/no-named-as-default
+import Conf, {ConfSchema} from '../source';
 
 const test = anyTest as TestInterface<{
 	config: Conf;
@@ -20,8 +21,8 @@ const test = anyTest as TestInterface<{
 const fixture = '🦄';
 
 test.beforeEach(t => {
-	t.context.config = new Conf({ cwd: tempy.directory() });
-	t.context.configWithoutDotNotation = new Conf({ cwd: tempy.directory(), accessPropertiesByDotNotation: false });
+	t.context.config = new Conf({cwd: tempy.directory()});
+	t.context.configWithoutDotNotation = new Conf({cwd: tempy.directory(), accessPropertiesByDotNotation: false});
 });
 
 test('.get()', t => {
@@ -51,51 +52,50 @@ test('.set() - with object', t => {
 	});
 	t.is(t.context.config.get('foo1'), 'bar1');
 	t.is(t.context.config.get('foo2'), 'bar2');
-	t.deepEqual(t.context.config.get('baz'), { boo: 'foo', foo: { bar: 'baz' } });
+	t.deepEqual(t.context.config.get('baz'), {boo: 'foo', foo: {bar: 'baz'}});
 	t.is(t.context.config.get('baz.boo'), 'foo');
-	t.deepEqual(t.context.config.get('baz.foo'), { bar: 'baz' });
+	t.deepEqual(t.context.config.get('baz.foo'), {bar: 'baz'});
 	t.is(t.context.config.get('baz.foo.bar'), 'baz');
 });
 
 test('.set() - with undefined', t => {
 	t.throws(() => {
 		t.context.config.set('foo', undefined);
-	}, 'Use `delete()` to clear values');
+	}, {message: 'Use `delete()` to clear values'});
 });
 
 test('.set() - with unsupported values', t => {
 	t.throws(() => {
-		t.context.config.set('a', () => { });
-	}, /not supported by JSON/);
+		t.context.config.set('a', () => {});
+	}, {message: /not supported by JSON/});
 
 	t.throws(() => {
 		t.context.config.set('a', Symbol('a'));
-	}, /not supported by JSON/);
+	}, {message: /not supported by JSON/});
 
 	t.throws(() => {
 		t.context.config.set({
 			a: undefined
 		});
-	}, /not supported by JSON/);
+	}, {message: /not supported by JSON/});
 
 	t.throws(() => {
 		t.context.config.set({
 			a: () => { }
 		});
-	}, /not supported by JSON/);
+	}, {message: /not supported by JSON/});
 
 	t.throws(() => {
 		t.context.config.set({
 			a: Symbol('a')
 		});
-	}, /not supported by JSON/);
+	}, {message: /not supported by JSON/});
 });
 
 test('.set() - invalid key', t => {
 	t.throws(() => {
-		// @ts-ignore
 		t.context.config.set(1, 'unicorn');
-	}, 'Expected `key` to be of type `string` or `object`, got number');
+	}, {message: 'Expected `key` to be of type `string` or `object`, got number'});
 });
 
 test('.has()', t => {
@@ -141,7 +141,7 @@ test('.reset()', t => {
 });
 
 test('.delete()', t => {
-	const { config } = t.context;
+	const {config} = t.context;
 	config.set('foo', 'bar');
 	config.set('baz.boo', true);
 	config.set('baz.foo.bar', 'baz');
@@ -150,9 +150,9 @@ test('.delete()', t => {
 	config.delete('baz.boo');
 	t.not(config.get('baz.boo'), true);
 	config.delete('baz.foo');
-	t.not(config.get('baz.foo'), { bar: 'baz' });
-	config.set('foo.bar.baz', { awesome: 'icecream' });
-	config.set('foo.bar.zoo', { awesome: 'redpanda' });
+	t.not(config.get('baz.foo'), {bar: 'baz'});
+	config.set('foo.bar.baz', {awesome: 'icecream'});
+	config.set('foo.bar.zoo', {awesome: 'redpanda'});
 	config.delete('foo.bar.baz');
 	t.is(config.get('foo.bar.zoo.awesome'), 'redpanda');
 });
@@ -213,7 +213,7 @@ test('no `suffix` option', t => {
 test('with `suffix` option set to empty string', t => {
 	const projectSuffix = '';
 	const projectName = 'conf-temp1-project';
-	const config = new Conf({ projectSuffix, projectName });
+	const config = new Conf({projectSuffix, projectName});
 	const configPathSegments = config.path.split(path.sep);
 	const configRootIndex = configPathSegments.findIndex(segment => segment === projectName);
 	t.true(configRootIndex >= 0 && configRootIndex < configPathSegments.length);
@@ -222,7 +222,7 @@ test('with `suffix` option set to empty string', t => {
 test('with `projectSuffix` option set to non-empty string', t => {
 	const projectSuffix = 'new-projectSuffix';
 	const projectName = 'conf-temp2-project';
-	const config = new Conf({ projectSuffix, projectName });
+	const config = new Conf({projectSuffix, projectName});
 	const configPathSegments = config.path.split(path.sep);
 	const expectedRootName = `${projectName}-${projectSuffix}`;
 	const configRootIndex = configPathSegments.findIndex(segment => segment === expectedRootName);
@@ -254,7 +254,7 @@ test('`fileExtension` option = empty string', t => {
 test('`serialize` and `deserialize` options', t => {
 	t.plan(4);
 	const serialized = `foo:${fixture}`;
-	const deserialized = { foo: fixture };
+	const deserialized = {foo: fixture};
 	const serialize = (value: unknown): string => {
 		t.is(value, deserialized);
 		return serialized;
@@ -277,19 +277,19 @@ test('`serialize` and `deserialize` options', t => {
 
 test('`projectName` option', t => {
 	const projectName = 'conf-fixture-project-name';
-	const config = new Conf({ projectName });
+	const config = new Conf({projectName});
 	t.is(config.get('foo'), undefined);
 	config.set('foo', fixture);
 	t.is(config.get('foo'), fixture);
 	t.true(config.path.includes(projectName));
-	del.sync(config.path, { force: true });
+	del.sync(config.path, {force: true});
 });
 
 test('ensure `.store` is always an object', t => {
 	const cwd = tempy.directory();
-	const config = new Conf({ cwd });
+	const config = new Conf({cwd});
 
-	del.sync(cwd, { force: true });
+	del.sync(cwd, {force: true});
 
 	t.notThrows(() => {
 		config.get('foo');
@@ -312,19 +312,19 @@ test('automatic `projectName` inference', t => {
 	config.set('foo', fixture);
 	t.is(config.get('foo'), fixture);
 	t.true(config.path.includes('conf'));
-	del.sync(config.path, { force: true });
+	del.sync(config.path, {force: true});
 });
 
 test('`cwd` option overrides `projectName` option', t => {
 	const cwd = tempy.directory();
 
 	t.notThrows(() => {
-		const config: Conf = new Conf({ cwd, projectName: '' });
+		const config: Conf = new Conf({cwd, projectName: ''});
 		t.true(config.path.startsWith(cwd));
 		t.is(config.get('foo'), undefined);
 		config.set('foo', fixture);
 		t.is(config.get('foo'), fixture);
-		del.sync(config.path, { force: true });
+		del.sync(config.path, {force: true});
 	});
 });
 
@@ -334,8 +334,8 @@ test('safely handle missing package.json', t => {
 
 	let config: Conf;
 	t.notThrows(() => {
-		config = new Conf({ projectName: 'conf-fixture-project-name' });
-		del.sync(config.path, { force: true });
+		config = new Conf({projectName: 'conf-fixture-project-name'});
+		del.sync(config.path, {force: true});
 	});
 
 	pkgUp.sync = pkgUpSyncOrig;
@@ -347,7 +347,7 @@ test('handle `cwd` being set and `projectName` not being set', t => {
 
 	let config: Conf;
 	t.notThrows(() => {
-		config = new Conf({ cwd: 'conf-fixture-cwd' });
+		config = new Conf({cwd: 'conf-fixture-cwd'});
 		del.sync(path.dirname(config.path));
 	});
 
@@ -361,15 +361,14 @@ test('fallback to cwd if `module.filename` is `null`', t => {
 	clearModule('.');
 
 	t.notThrows(() => {
-		const Conf = require('../dist').default;
-		const config: Conf = new Conf({ cwd: 'conf-fixture-fallback-module-filename-null' });
+		const config: Conf = new Conf({cwd: 'conf-fixture-fallback-module-filename-null'});
 		del.sync(path.dirname(config.path));
 		module.filename = preservedFilename;
 	});
 });
 
 test('encryption', t => {
-	const config = new Conf({ cwd: tempy.directory(), encryptionKey: 'abc123' });
+	const config = new Conf({cwd: tempy.directory(), encryptionKey: 'abc123'});
 	t.is(config.get('foo'), undefined);
 	t.is(config.get('foo', '🐴'), '🐴');
 	config.set('foo', fixture);
@@ -381,24 +380,24 @@ test('encryption', t => {
 test('encryption - upgrade', t => {
 	const cwd = tempy.directory();
 
-	const before = new Conf({ cwd });
+	const before = new Conf({cwd});
 	before.set('foo', fixture);
 	t.is(before.get('foo'), fixture);
 
-	const after = new Conf({ cwd, encryptionKey: 'abc123' });
+	const after = new Conf({cwd, encryptionKey: 'abc123'});
 	t.is(after.get('foo'), fixture);
 });
 
 test('encryption - corrupt file', t => {
 	const cwd = tempy.directory();
 
-	const before = new Conf({ cwd, encryptionKey: 'abc123' });
+	const before = new Conf({cwd, encryptionKey: 'abc123'});
 	before.set('foo', fixture);
 	t.is(before.get('foo'), fixture);
 
 	fs.appendFileSync(path.join(cwd, 'config.json'), 'corrupt file');
 
-	const after = new Conf({ cwd, encryptionKey: 'abc123' });
+	const after = new Conf({cwd, encryptionKey: 'abc123'});
 	t.is(after.get('foo'), undefined);
 });
 
@@ -410,11 +409,11 @@ test('decryption - migration to initialization vector', t => {
 		configName: 'config-encrypted-with-conf-4-1-0'
 	});
 
-	t.deepEqual(config.store, { unicorn: '🦄' });
+	t.deepEqual(config.store, {unicorn: '🦄'});
 });
 
 test('onDidChange()', t => {
-	const { config } = t.context;
+	const {config} = t.context;
 
 	t.plan(8);
 
@@ -460,23 +459,23 @@ test('onDidChange()', t => {
 });
 
 test('onDidAnyChange()', t => {
-	const { config } = t.context;
+	const {config} = t.context;
 
 	t.plan(8);
 
 	const checkFoo = (newValue: unknown, oldValue: unknown): void => {
-		t.deepEqual(newValue, { foo: '🐴' });
-		t.deepEqual(oldValue, { foo: fixture });
+		t.deepEqual(newValue, {foo: '🐴'});
+		t.deepEqual(oldValue, {foo: fixture});
 	};
 
 	const checkBaz = (newValue: unknown, oldValue: unknown): void => {
 		t.deepEqual(newValue, {
 			foo: fixture,
-			baz: { boo: '🐴' }
+			baz: {boo: '🐴'}
 		});
 		t.deepEqual(oldValue, {
 			foo: fixture,
-			baz: { boo: fixture }
+			baz: {boo: fixture}
 		});
 	};
 
@@ -495,21 +494,21 @@ test('onDidAnyChange()', t => {
 	const checkUndefined = (newValue: unknown, oldValue: unknown): void => {
 		t.deepEqual(oldValue, {
 			foo: '🦄',
-			baz: { boo: '🦄' }
+			baz: {boo: '🦄'}
 		});
 
 		t.deepEqual(newValue, {
-			baz: { boo: fixture }
+			baz: {boo: fixture}
 		});
 	};
 
 	const checkSet = (newValue: unknown, oldValue: unknown): void => {
 		t.deepEqual(oldValue, {
-			baz: { boo: fixture }
+			baz: {boo: fixture}
 		});
 
 		t.deepEqual(newValue, {
-			baz: { boo: '🦄' },
+			baz: {boo: '🦄'},
 			foo: '🐴'
 		});
 	};
@@ -539,30 +538,30 @@ test('doesn\'t write to disk upon instanciation if and only if the store didn\'t
 });
 
 test('`clearInvalidConfig` option - invalid data', t => {
-	const config = new Conf({ cwd: tempy.directory(), clearInvalidConfig: false });
+	const config = new Conf({cwd: tempy.directory(), clearInvalidConfig: false});
 	fs.writeFileSync(config.path, '🦄');
 
 	t.throws(() => {
 		// eslint-disable-next-line @typescript-eslint/no-unused-expressions
 		config.store;
-	}, { instanceOf: SyntaxError });
+	}, {instanceOf: SyntaxError});
 });
 
 test('`clearInvalidConfig` option - valid data', t => {
-	const config = new Conf({ cwd: tempy.directory(), clearInvalidConfig: false });
+	const config = new Conf({cwd: tempy.directory(), clearInvalidConfig: false});
 	config.set('foo', 'bar');
-	t.deepEqual(config.store, { foo: 'bar' });
+	t.deepEqual(config.store, {foo: 'bar'});
 });
 
 test('schema - should be an object', t => {
 	const schema: any = 'object';
 	t.throws(() => {
-		new Conf({ cwd: tempy.directory(), schema });
-	}, 'The `schema` option must be an object.');
+		new Conf({cwd: tempy.directory(), schema});
+	}, {message: 'The `schema` option must be an object.'});
 });
 
 test('schema - valid set', t => {
-	const schema = {
+	const schema: ConfSchema<{ foo: { bar: number; foobar: number } }> = {
 		foo: {
 			type: 'object',
 			properties: {
@@ -576,26 +575,28 @@ test('schema - valid set', t => {
 			}
 		}
 	};
-	const config = new Conf({ cwd: tempy.directory(), schema });
+	const config = new Conf({cwd: tempy.directory(), schema});
 	t.notThrows(() => {
-		config.set('foo', { bar: 1, foobar: 2 });
+		config.set('foo', {bar: 1, foobar: 2});
 	});
 });
 
 test('schema - one violation', t => {
-	const schema = {
-		foo: {
-			type: 'string'
+	const config = new Conf({
+		cwd: tempy.directory(),
+		schema: {
+			foo: {
+				type: 'string'
+			}
 		}
-	};
-	const config = new Conf({ cwd: tempy.directory(), schema });
+	});
 	t.throws(() => {
 		config.set('foo', 1);
-	}, 'Config schema violation: `foo` should be string');
+	}, {message: 'Config schema violation: `foo` should be string'});
 });
 
 test('schema - multiple violations', t => {
-	const schema = {
+	const schema: ConfSchema<{ foo: { bar: number; foobar: number } }> = {
 		foo: {
 			type: 'object',
 			properties: {
@@ -609,14 +610,14 @@ test('schema - multiple violations', t => {
 			}
 		}
 	};
-	const config = new Conf({ cwd: tempy.directory(), schema });
+	const config = new Conf({cwd: tempy.directory(), schema});
 	t.throws(() => {
-		config.set('foo', { bar: '1', foobar: 101 });
-	}, 'Config schema violation: `foo.bar` should be number; `foo.foobar` should be <= 100');
+		config.set('foo', {bar: '1', foobar: 101});
+	}, {message: 'Config schema violation: `foo.bar` should be number; `foo.foobar` should be <= 100'});
 });
 
 test('schema - complex schema', t => {
-	const schema = {
+	const schema: ConfSchema<{ foo: string; bar: number[] }> = {
 		foo: {
 			type: 'string',
 			maxLength: 3,
@@ -631,32 +632,32 @@ test('schema - complex schema', t => {
 			}
 		}
 	};
-	const config = new Conf({ cwd: tempy.directory(), schema });
+	const config = new Conf({cwd: tempy.directory(), schema});
 	t.throws(() => {
 		config.set('foo', 'abca');
-	}, 'Config schema violation: `foo` should NOT be longer than 3 characters; `foo` should match pattern "[def]+"');
+	}, {message: 'Config schema violation: `foo` should NOT be longer than 3 characters; `foo` should match pattern "[def]+"'});
 	t.throws(() => {
 		config.set('bar', [1, 1, 2, 'a']);
-	}, 'Config schema violation: `bar` should NOT have more than 3 items; `bar[3]` should be integer; `bar` should NOT have duplicate items (items ## 1 and 0 are identical)');
+	}, {message: 'Config schema violation: `bar` should NOT have more than 3 items; `bar[3]` should be integer; `bar` should NOT have duplicate items (items ## 1 and 0 are identical)'});
 });
 
 test('schema - invalid write to config file', t => {
-	const schema = {
+	const schema: ConfSchema<{ foo: string }> = {
 		foo: {
 			type: 'string'
 		}
 	};
 	const cwd = tempy.directory();
 
-	const config = new Conf({ cwd, schema });
-	fs.writeFileSync(path.join(cwd, 'config.json'), JSON.stringify({ foo: 1 }));
+	const config = new Conf({cwd, schema});
+	fs.writeFileSync(path.join(cwd, 'config.json'), JSON.stringify({foo: 1}));
 	t.throws(() => {
 		config.get('foo');
-	}, 'Config schema violation: `foo` should be string');
+	}, {message: 'Config schema violation: `foo` should be string'});
 });
 
 test('schema - default', t => {
-	const schema = {
+	const schema: ConfSchema<{ foo: string }> = {
 		foo: {
 			type: 'string',
 			default: 'bar'
@@ -670,7 +671,7 @@ test('schema - default', t => {
 });
 
 test('schema - Conf defaults overwrites schema default', t => {
-	const schema = {
+	const schema: ConfSchema<{ foo: string }> = {
 		foo: {
 			type: 'string',
 			default: 'bar'
@@ -687,7 +688,7 @@ test('schema - Conf defaults overwrites schema default', t => {
 });
 
 test('schema - validate Conf default', t => {
-	const schema = {
+	const schema: ConfSchema<{ foo: string }> = {
 		foo: {
 			type: 'string'
 		}
@@ -696,11 +697,14 @@ test('schema - validate Conf default', t => {
 		new Conf({
 			cwd: tempy.directory(),
 			defaults: {
+				// For our tests to fail and typescript to compile, we'll ignore this ts error.
+				// This error is not bad and means the package is well typed.
+				// @ts-ignore
 				foo: 1
 			},
 			schema
 		});
-	}, 'Config schema violation: `foo` should be string');
+	}, {message: 'Config schema violation: `foo` should be string'});
 });
 
 test('.get() - without dot notation', t => {
@@ -730,7 +734,7 @@ test('.set() - with object - without dot notation', t => {
 	});
 	t.is(t.context.configWithoutDotNotation.get('foo1'), 'bar1');
 	t.is(t.context.configWithoutDotNotation.get('foo2'), 'bar2');
-	t.deepEqual(t.context.configWithoutDotNotation.get('baz'), { boo: 'foo', foo: { bar: 'baz' } });
+	t.deepEqual(t.context.configWithoutDotNotation.get('baz'), {boo: 'foo', foo: {bar: 'baz'}});
 	t.is(t.context.configWithoutDotNotation.get('baz.boo'), undefined);
 	t.is(t.context.configWithoutDotNotation.get('baz.foo.bar'), undefined);
 });
@@ -744,7 +748,7 @@ test('.has() - without dot notation', t => {
 });
 
 test('.delete() - without dot notation', t => {
-	const { configWithoutDotNotation } = t.context;
+	const {configWithoutDotNotation} = t.context;
 	configWithoutDotNotation.set('foo', 'bar');
 	configWithoutDotNotation.set('baz.boo', true);
 	configWithoutDotNotation.set('baz.foo.bar', 'baz');
@@ -753,11 +757,11 @@ test('.delete() - without dot notation', t => {
 	configWithoutDotNotation.delete('baz.boo');
 	t.not(configWithoutDotNotation.get('baz.boo'), true);
 	configWithoutDotNotation.delete('baz.foo');
-	t.not(configWithoutDotNotation.get('baz.foo'), { bar: 'baz' });
-	configWithoutDotNotation.set('foo.bar.baz', { awesome: 'icecream' });
-	configWithoutDotNotation.set('foo.bar.zoo', { awesome: 'redpanda' });
+	t.not(configWithoutDotNotation.get('baz.foo'), {bar: 'baz'});
+	configWithoutDotNotation.set('foo.bar.baz', {awesome: 'icecream'});
+	configWithoutDotNotation.set('foo.bar.zoo', {awesome: 'redpanda'});
 	configWithoutDotNotation.delete('foo.bar.baz');
-	t.deepEqual(configWithoutDotNotation.get('foo.bar.zoo'), { awesome: 'redpanda' });
+	t.deepEqual(configWithoutDotNotation.get('foo.bar.zoo'), {awesome: 'redpanda'});
 });
 
 test('`watch` option watches for config file changes by another process', async t => {
@@ -767,8 +771,8 @@ test('`watch` option watches for config file changes by another process', async 
 	}
 
 	const cwd = tempy.directory();
-	const conf1 = new Conf({ cwd, watch: true });
-	const conf2 = new Conf({ cwd });
+	const conf1 = new Conf({cwd, watch: true});
+	const conf2 = new Conf({cwd});
 	conf1.set('foo', '👾');
 
 	t.plan(4);
@@ -787,7 +791,7 @@ test('`watch` option watches for config file changes by another process', async 
 		conf2.set('foo', '🐴');
 	})();
 
-	const { _events } = conf1;
+	const {events: _events} = conf1;
 
 	await pEvent(_events, 'change');
 });
@@ -800,7 +804,7 @@ test('`watch` option watches for config file changes by file write', async t => 
 	}
 
 	const cwd = tempy.directory();
-	const conf = new Conf({ cwd, watch: true });
+	const conf = new Conf({cwd, watch: true});
 	conf.set('foo', '🐴');
 
 	t.plan(2);
@@ -814,18 +818,18 @@ test('`watch` option watches for config file changes by file write', async t => 
 
 	(async () => {
 		await delay(50);
-		fs.writeFileSync(path.join(cwd, 'config.json'), JSON.stringify({ foo: '🦄' }));
+		fs.writeFileSync(path.join(cwd, 'config.json'), JSON.stringify({foo: '🦄'}));
 	})();
 
-	const { _events } = conf;
+	const {events} = conf;
 
-	await pEvent(_events, 'change');
+	await pEvent(events, 'change');
 });
 
 test('migrations - should save the project version as the initial migrated version', t => {
 	const cwd = tempy.directory();
 
-	const conf = new Conf({ cwd, projectVersion: '0.0.2', migrations: {} });
+	const conf = new Conf({cwd, projectVersion: '0.0.2', migrations: {}});
 
 	t.is(conf.get('__internal__.migrations.version'), '0.0.2');
 });
@@ -839,11 +843,11 @@ test('migrations - should save the project version when a migration occurs', t =
 		}
 	};
 
-	const conf = new Conf({ cwd, projectVersion: '0.0.2', migrations });
+	const conf = new Conf({cwd, projectVersion: '0.0.2', migrations});
 
 	t.is(conf.get('__internal__.migrations.version'), '0.0.2');
 
-	const conf2 = new Conf({ cwd, projectVersion: '0.0.4', migrations });
+	const conf2 = new Conf({cwd, projectVersion: '0.0.4', migrations});
 
 	t.is(conf2.get('__internal__.migrations.version'), '0.0.4');
 	t.is(conf2.get('foo'), 'cool stuff');
@@ -858,11 +862,11 @@ test('migrations - should NOT run the migration when the version doesn\'t change
 		}
 	};
 
-	const conf = new Conf({ cwd, projectVersion: '0.0.2', migrations });
+	const conf = new Conf({cwd, projectVersion: '0.0.2', migrations});
 	t.is(conf.get('__internal__.migrations.version'), '0.0.2');
 	t.false(conf.has('foo'));
 
-	const conf2 = new Conf({ cwd, projectVersion: '0.0.2', migrations });
+	const conf2 = new Conf({cwd, projectVersion: '0.0.2', migrations});
 
 	t.is(conf2.get('__internal__.migrations.version'), '0.0.2');
 	t.false(conf2.has('foo'));
@@ -877,11 +881,11 @@ test('migrations - should run the migration when the version changes', t => {
 		}
 	};
 
-	const conf = new Conf({ cwd, projectVersion: '0.0.2', migrations });
+	const conf = new Conf({cwd, projectVersion: '0.0.2', migrations});
 	t.is(conf.get('__internal__.migrations.version'), '0.0.2');
 	t.false(conf.has('foo'));
 
-	const conf2 = new Conf({ cwd, projectVersion: '1.1.0', migrations });
+	const conf2 = new Conf({cwd, projectVersion: '1.1.0', migrations});
 
 	t.is(conf2.get('__internal__.migrations.version'), '1.1.0');
 	t.true(conf2.has('foo'));
@@ -896,7 +900,7 @@ test('migrations - should run the migration when the version uses semver compari
 		}
 	};
 
-	const conf = new Conf({ cwd, projectVersion: '1.0.2', migrations });
+	const conf = new Conf({cwd, projectVersion: '1.0.2', migrations});
 	t.is(conf.get('__internal__.migrations.version'), '1.0.2');
 	t.is(conf.get('foo'), 'cool stuff');
 });
@@ -912,11 +916,11 @@ test('migrations - should run the migration when the version uses multiple semve
 		}
 	};
 
-	const conf = new Conf({ cwd, projectVersion: '1.0.2', migrations });
+	const conf = new Conf({cwd, projectVersion: '1.0.2', migrations});
 	t.is(conf.get('__internal__.migrations.version'), '1.0.2');
 	t.is(conf.get('foo'), 'cool stuff');
 
-	const conf2 = new Conf({ cwd, projectVersion: '2.0.1', migrations });
+	const conf2 = new Conf({cwd, projectVersion: '2.0.1', migrations});
 	t.is(conf2.get('__internal__.migrations.version'), '2.0.1');
 	t.is(conf2.get('foo'), 'modern cool stuff');
 });
@@ -937,7 +941,7 @@ test('migrations - should run all valid migrations when the version uses multipl
 		}
 	};
 
-	const conf = new Conf({ cwd, projectVersion: '2.4.0', migrations });
+	const conf = new Conf({cwd, projectVersion: '2.4.0', migrations});
 	t.is(conf.get('__internal__.migrations.version'), '2.4.0');
 	t.is(conf.get('foo'), 'cool stuff');
 	t.is(conf.get('medium'), 'yes');
@@ -961,7 +965,7 @@ test('migrations - should cleanup migrations with non-numeric values', t => {
 		}
 	};
 
-	const conf = new Conf({ cwd, projectVersion: '2.4.0', migrations });
+	const conf = new Conf({cwd, projectVersion: '2.4.0', migrations});
 	t.is(conf.get('__internal__.migrations.version'), '2.4.0');
 	t.is(conf.get('foo'), 'cool stuff');
 	t.is(conf.get('medium'), 'yes');
@@ -981,14 +985,14 @@ test('migrations - should infer the applicationVersion from the package.json whe
 	});
 
 	t.false(conf.has('foo'));
-	t.is(conf.get('__internal__.migrations.version'), require('../package.json').version);
+	t.is(conf.get('__internal__.migrations.version'), packageVersion);
 });
 
 test('migrations - should NOT throw an error when project version is unspecified but there are no migrations', t => {
 	const cwd = tempy.directory();
 
 	t.notThrows(() => {
-		const conf = new Conf({ cwd });
+		const conf = new Conf({cwd});
 		conf.clear();
 	});
 });
@@ -996,7 +1000,7 @@ test('migrations - should NOT throw an error when project version is unspecified
 test('migrations - should not create the previous migration key if the migrations aren\'t needed', t => {
 	const cwd = tempy.directory();
 
-	const conf = new Conf({ cwd });
+	const conf = new Conf({cwd});
 	t.false(conf.has('__internal__.migrations.version'));
 });
 
@@ -1023,11 +1027,11 @@ test('migrations error handling - should rollback changes if a migration failed'
 		}
 	};
 
-	let conf = new Conf({ cwd, projectVersion: '1.0.0', migrations: passingMigrations });
+	let conf = new Conf({cwd, projectVersion: '1.0.0', migrations: passingMigrations});
 
 	t.throws(() => {
-		conf = new Conf({ cwd, projectVersion: '1.0.2', migrations: failingMigrations });
-	}, /throw the migration and rollback/);
+		conf = new Conf({cwd, projectVersion: '1.0.2', migrations: failingMigrations});
+	}, {message: /throw the migration and rollback/});
 
 	t.is(conf.get('__internal__.migrations.version'), '1.0.0');
 	t.true(conf.has('foo'));
@@ -1037,17 +1041,17 @@ test('migrations error handling - should rollback changes if a migration failed'
 test('__internal__ keys - should not be accessible by the user', t => {
 	const cwd = tempy.directory();
 
-	const conf = new Conf({ cwd });
+	const conf = new Conf({cwd});
 
 	t.throws(() => {
 		conf.set('__internal__.you-shall', 'not-pass');
-	}, /Please don't use the __internal__ key/);
+	}, {message: /Please don't use the __internal__ key/});
 });
 
 test('__internal__ keys - should not be accessible by the user even without dot notation', t => {
 	const cwd = tempy.directory();
 
-	const conf = new Conf({ cwd, accessPropertiesByDotNotation: false });
+	const conf = new Conf({cwd, accessPropertiesByDotNotation: false});
 
 	t.throws(() => {
 		conf.set({
@@ -1055,13 +1059,13 @@ test('__internal__ keys - should not be accessible by the user even without dot 
 				'you-shall': 'not-pass'
 			}
 		});
-	}, /Please don't use the __internal__ key/);
+	}, {message: /Please don't use the __internal__ key/});
 });
 
 test('__internal__ keys - should only match specific "__internal__" entry', t => {
 	const cwd = tempy.directory();
 
-	const conf = new Conf({ cwd });
+	const conf = new Conf({cwd});
 
 	t.notThrows(() => {
 		conf.set('__internal__foo.you-shall', 'not-pass');
